+++ src/so101_bench/file_utils.py
@@ -1,0 +1,14 @@
+import yaml
+from pathlib import Path
+
+
+def load_yaml(file_path: Path) -> dict:
+    """Load YAML file."""
+    with open(file_path, 'r') as f:
+        return yaml.safe_load(f)
+
+
+def save_yaml(data: dict, file_path: Path):
+    """Save data to YAML file."""
+    with open(file_path, 'w') as f:
+        yaml.dump(data, f, default_flow_style=False, indent=2)
